+++ src/notifications/hub.ts
@@ -1,6 +1,5 @@
 import * as Request from "../common/api/request";
-import { links, StandardParams } from "../common/api";
-import { connectToSocket } from "../common/api/websocket";
+import { links, StandardParams, connectToSocket } from "../common/api";
 import { Notification } from "./notification";
 
 /**
@@ -63,15 +62,11 @@ export type HubTopic =
   | "dns.zone.created"
   | "dns.zone.verified"
   | "dns.zone.reconfigured"
-  | "dns.zone.records.reconfigured"
   | "dns.zone.certificate.ready"
 
   // dns zone record
   | "dns.zone.record.state.changed"
-
-  // employees
-  | "employee.state.changed"
-  | "employee.error"
+  | "dns.zone.records.reconfigured"
 
   // environments
   | "environment.started"
@@ -85,10 +80,10 @@ export type HubTopic =
   | "environment.services.lb.ip.acquired"
 
   // hub
-  | "hub.updated"
+  | "hub.activity.new"
   | "hub.state.changed"
   | "hub.error"
-  | "hub.activity.new"
+  | "hub.updated"
 
   // hub api keys
   | "hub.api_key.created"
@@ -100,6 +95,7 @@ export type HubTopic =
   | "hub.membership.state.changed"
   | "hub.membership.error"
   | "hub.membership.updated"
+  | "hub.membership.new"
 
   // images
   | "image.created"
@@ -118,8 +114,8 @@ export type HubTopic =
   | "infrastructure.ips.assignment.error"
 
   // infrastructure ips pool
-  | "infrastructure.ips.pool.state.changed"
-  | "infrastructure.ips.pool.error"
+  | "ips_pool.state.changed"
+  | "ips_pool.error"
 
   // infrastructure server
   | "infrastructure.server.state.changed"
@@ -130,7 +126,23 @@ export type HubTopic =
   // jobs
   | "job.created"
   | "job.state.changed"
-  | "job.created"
+
+  // pipeline
+  | "pipeline.state.changed"
+  | "pipeline.error"
+  | "pipeline.updated"
+  | "pipeline.created"
+
+  // pipeline key
+  | "pipeline.key.state.changed"
+  | "pipeline.key.error"
+  | "pipeline.key.updated"
+  | "pipeline.key.created"
+
+  // pipeline run
+  | "pipeline.run.state.changed"
+  | "pipeline.run.error"
+  | "pipeline.run.created"
 
   // sdn
   | "sdn.network.created"
@@ -146,34 +158,14 @@ export type HubTopic =
   // stack
   | "stack.state.changed"
   | "stack.error"
-
-  // stack
   | "stack.created"
-  | "stack.error"
-  | "stack.state.changed"
   | "stack.updated"
 
-  // builds
+  // stack builds
   | "stack.build.created"
   | "stack.build.state.changed"
   | "stack.build.error"
-  | "stack.build.deployed"
-
-  // pipelines
-  // pipeline
-  | "pipeline.state.changed"
-  | "pipeline.error"
-  | "pipeline.updated"
-  | "pipeline.created"
-  // key
-  | "pipeline.key.state.changed"
-  | "pipeline.key.error"
-  | "pipeline.key.updated"
-  | "pipeline.key.created"
-  // run
-  | "pipeline.run.state.changed"
-  | "pipeline.run.error"
-  | "pipeline.run.created";
+  | "stack.build.deployed";
 
 export type HubNotification = Notification<HubTopic>;