+++ src/resources/stacks/builds/tasks/build.ts
@@ -4,22 +4,62 @@ import { ResourceId, Task, CreatedTask } from "../../../../common/structs";
 
 export type BuildAction = "deploy" | "delete" | "generate";
 
-export interface DeployParams {
+export interface DeployContents {
+  /** The id of the environment to update with the stack build */
   environment_id: ResourceId;
-  update_configs: boolean;
-  redeploy: boolean;
+  /**
+   * Optional update object used to specify specific params to update from
+   * the stack build.
+   */
+  update?: DeployContentsUpdate;
 }
 
-export async function deployBuild(
+/**
+ * The update interface used for the `update` key inside
+ * of the `DeployContents` interface
+ */
+export interface DeployContentsUpdate {
+  /**
+   * This is a map of the container names to update within the
+   * environment. The map follows the format of:
+   * ```
+   *  {
+   *    container_id: {
+   *      reimage: true,
+   *      reconfigure: true,
+   *    }
+   *  }
+   * ```
+   */
+  containers: Record<ResourceId, DeployUpdateContainers>;
+}
+
+export interface DeployUpdateContainers {
+  /**
+   * If set to true the container will be reimaged with the image specified in
+   * the stack build
+   */
+  reimage: boolean;
+  /**
+   * If set to true the container will use the new `config` settings specified in
+   * the stack build
+   */
+  reconfigure: boolean;
+}
+
+export async function deploy(
   params: StandardParams & {
     id: ResourceId;
     stackId: ResourceId;
-    value: DeployParams;
+    contents: DeployContents;
   },
 ) {
-  return task<DeployParams>({
+  return task<DeployContents>({
     ...params,
-    value: { action: "deploy", contents: params.value },
+    value: {
+      action: "deploy",
+      contents: params.contents,
+    },
   });
 }
 
@@ -29,7 +69,7 @@ export async function generate(
     stackId: ResourceId;
   },
 ) {
-  return task<DeployParams>({
+  return task({
     ...params,
     value: { action: "generate" },
   });
@@ -43,10 +83,7 @@ export async function remove(
 ) {
   return Request.deleteRequest<CreatedTask<"delete">>({
     ...params,
-    target: links
-      .stacks()
-      .builds(params.stackId)
-      .single(params.id),
+    target: links.stacks().builds(params.stackId).single(params.id),
   });
 }
 
@@ -59,9 +96,6 @@ export async function task<K = {}>(
 ) {
   return Request.postRequest<CreatedTask<BuildAction, K>>({
     ...params,
-    target: links
-      .stacks()
-      .builds(params.stackId)
-      .tasks(params.id),
+    target: links.stacks().builds(params.stackId).tasks(params.id),
   });
 }